requirements.txt
streamlit==1.26.0
plotly==5.15.0
pandas==2.0.3
numpy==1.25.0
folium==0.14.0
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.13.0
Pillow==9.5.0